+++ lib/base/base.php
@@ -5,11 +5,11 @@
  * @author      Martin Gleiß
  * @copyright   2012 - 2015
  * @license     GPL [http://www.gnu.de]
- * ----------------------------------------------------------------------------- 
+ * -----------------------------------------------------------------------------
  */
- 
- 
-    // get config-variables 
+
+
+    // get config-variables
     require_once '../../lib/includes.php';
 
 	@header('Cache-Control: public, must-revalidate');
@@ -20,8 +20,8 @@
 /**
  * -----------------------------------------------------------------------------
  * J A V A S C R I P T   D Y N A M I C   E X T E N T I O N S
- * ----------------------------------------------------------------------------- 
- */ 
+ * -----------------------------------------------------------------------------
+ */
 
 
 // ----- number ----------------------------------------------------------------
@@ -31,11 +31,11 @@
 	*/
 	Number.prototype.transUnit = function(unit) {
 		<?php echo 'formats  = ' . trans('format', '', 'obj') . ';'; ?>
-		
+
 		if (unit == '') {
 			return this;
 		}
-		
+
 		var fmt = formats[unit.toLowerCase()];
 		if (fmt !== undefined)
 			return printf(fmt, this);
@@ -51,7 +51,7 @@
 	Number.prototype.transInt = function() {
 		return this.transUnit('int');
 	};
-	
+
   /**
 	* transforms a number to float
 	*/
@@ -66,25 +66,58 @@
     * transforms a date to date
     */
 	Date.prototype.transUnit = function(unit) {
-		<?php echo 'formats  = ' . trans('format', '', 'obj') . ';'; ?>
-		
+		<?php echo 'formats = ' . trans('format', '', 'obj') . ';'; ?>
+		<?php echo 'month = ' . trans('month', '', '') . ';'; ?>
+		<?php echo 'shortmonth = ' . trans('shortmonth', '', '') . ';'; ?>
+		<?php echo 'weekday = ' . trans('weekday', '', '') . ';'; ?>
+		<?php echo 'shortday = ' . trans('shortday', '', '') . ';'; ?>
+
 		if (unit == '') {
 			return this;
 		}
 
 		var ret = formats[unit.toLowerCase()];
-		if (ret == undefined) {
-			return this.toString() + ' ' + unit;
-		}
+		if (ret == undefined)
+			ret = unit;
+// 259200000
 
-		ret = ret.replace('d', ('0' + this.getDate()).slice(-2));
-		ret = ret.replace('m', ('0' + (this.getMonth() + 1)).slice(-2));
-		ret = ret.replace('y', this.getFullYear().toString().substr(2, 2));
-		ret = ret.replace('Y', this.getFullYear());
+		var val = this;
 
-		ret = ret.replace('H', this.getHours());
-		ret = ret.replace('i', ('0' + this.getMinutes()).slice(-2));
-		ret = ret.replace('s', ('0' + this.getSeconds()).slice(-2));
+		// iterate over each character
+		ret = ret.replace(/./g, function(char) {
+			switch (char) {
+				case 'l':
+					return weekday[(val.getDay() + 1) % 7];
+				case 'D':
+					return shortday[(val.getDay() + 1) % 7];
+				case 'd':
+					return ('0' + val.getDate()).slice(-2);
+				case 'j':
+					return val.getDate();
+				case 'F':
+					return month[val.getMonth()];
+				case 'M':
+					return shortmonth[val.getMonth()];
+				case 'm':
+					return ('0' + (val.getMonth() + 1)).slice(-2);
+				case 'n':
+					return val.getMonth();
+				case 'y':
+					return val.getFullYear().toString().substr(2, 2);
+				case 'Y':
+					return val.getFullYear();
+				case 'H':
+					return ('0' + val.getHours()).slice(-2);
+				case 'G':
+					return val.getHours();
+				case 'i':
+					return ('0' + val.getMinutes()).slice(-2);
+				case 's':
+					return ('0' + val.getSeconds()).slice(-2);
+				default:
+					return char;
+			};
+		});
 
 		return ret;
 	};
@@ -110,20 +143,19 @@
 		return this.transUnit('long');
     };
 
-
 /**
  * -----------------------------------------------------------------------------
  * P L O T   D Y N A M I C   E X T E N T I O N S
- * ----------------------------------------------------------------------------- 
- */ 
+ * -----------------------------------------------------------------------------
+ */
 
   /**
     * Plots
     */
 	$(function () {
-		
+
 		Highcharts.setOptions({
-			lang: { 
+			lang: {
 				months: <?php echo trans('month') ?>,
 				shortMonths: <?php echo trans('shortmonth') ?>,
 				weekdays: <?php echo trans('weekday') ?>,
@@ -136,7 +168,7 @@
 				}
 			}
 	    });
-	
+
 	});